feat(server): contract modified should navigate to constructor
If the constructor is available the definition node should be the constructor rather than the entire contract. If there is no constructor fall back to the contract definition.

The opposite applies for references, so that find references from a constructur should show subcontracts modifiers. Previously the constructor was ignored as it has no name. The name is now set to the parent contract if available.