feat(completion): support import completion on `"`
We ignore completion on strings outside of import statements, but we can
give immediate completions within import statements now we trigger on
strings.

To support this implies modifying the completion list as the user goes
from `""` to `"."` to `"./"`. We now separate out the displayed text of
the completion from the insertion text so the user will see the
expected unchanging list, but the insertions will change under the
surface.

There had to be a fix for the "_;" inserted after "." that is added to
aid ast parsing. Basically we remove it again.

Fixes #52.