feat(completion): add direct imports from index
Support direct imports based on the validation index.

We scan the project for sol files on startup, this pulls in sol
files under `node_modules`. On an empty import (i.e. `import "";`), we
complete based on relative files in the same directory, and the names of
any packages in `node_modules` that contain sol files in the validation
index.

Direct imports triggered on a slash within one of the `node_module`
packages (i.e. `import "@openzeppelin/contracts/";`) get completed
with each known sol file, so `@openzeppelin` would get completed
with a list of the known sol file paths within `@openzeppelin` from
the validation list.

For partial paths (re-trigger with ctrl-space), to correctly give a list
of completions we switch to doing a replace, this gibes us proper prefix
highlighting in vscode (including '@' for those npm org imports).