chore: bundle the extension
Adds a new package.json scripts command `package`, that bundles the
extension code and creates a `.vsix` file for release.

Esbuild is used and added as a dev dependency for the bundling and
minification.

Bundling happens in place (to simplify things), the referenced files
in package.json are replaced with bundled and minified versions in the
client and server `out` directories.

The internal folder structure of the `vsix` file matches up with
the `/out` folders in the app during development, but this simplifies
getting debugging and integration working without editing the
package.json file during one of packaging, or debugging or
integration testing.

The `.vscodeignore` file is used to exclude unnecessary files from the
extension, and has been increased so that only key files as zipped in
the `vsix` file.

Resolves #67.