feat(analytics): base ga off of telemetry setting
Make the google analytics sending conditional on the `telemetryEnabled`
global setting in vscode.

This removes the UI message to opt in/out of telemetry, as we operate
off of the global setting.

The server commands that are tracked bu google analytics for performance
reasons has been expanded. A refactored `trackTiming` method wraps this
logic up for use in the server.

The language server, telemetry and analytics have been refactored to
separate out construction and initialization, with construction
happening on app start and initialization happening during the
`onInitialize` and `onInitialized` commands. This has the benefit of
ensure that each of the components always existing in serverState (so
less null checking).

No analytics are sent in development.

Related to #40.