fix: correct windows uri returns
Before returning a response from the server, convert the internal uri to
one that works for vscode on windows (if on windows).

This has been done for:

* onDefinition
* validation
* rename

The uri used in the background thread has to be converted to be os
specific before processing.

Increase the default mocha timeout to 5s to reduce number of transient fails.