chore: change telemetry attribute name
The rename from `hardhat` to `solidity` did not migrate the enable
telemetry value. Hence we need to reask our users, this is most easily
done by changing the name of the global state that records if we have
asked before, triggering a reask.

The old global state value is removed.